requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
graphql-core==3.2.6
    # via strawberry-graphql
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
libcst==1.7.0
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
packaging==25.0
    # via strawberry-graphql
pydantic==2.11.4
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.1
    # via rich
python-dateutil==2.9.0.post0
    # via strawberry-graphql
python-dotenv==1.1.0
    # via uvicorn
python-multipart==0.0.20
pyyaml==6.0.2
    # via
    #   libcst
    #   uvicorn
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.6
    # via fastapi-cli
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
starlette==0.46.2
    # via fastapi
strawberry-graphql==0.268.1
typer==0.15.4
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   sqlalchemy
    #   strawberry-graphql
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
uvicorn==0.34.2
    # via fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
